Exhaustive search of param space Stepping vs. Brute Force
I have created code to perform extensive testing (of the entire paramter space)
of the Stepping algorithm against a brute force algorithm. Currently, I have to
manually specify the range of the okutsu invariants (1-3 for example) as well
as the number of types and their respective number of levels.

The plan is to use it to test lots of possible okutsu invariant configurations
to try and find one where the Stepping algorithm fails.